backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0

# Database
psycopg[binary]==3.2.13

# Authentication
djangorestframework-simplejwt==5.3.1
webauthn==2.3.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# Validation
django-filter==23.5